fix(server): 更改Host部分数据类型
memory Integer -> Long
disk Integer -> Long

+++ server/src/main/java/org/pi/server/model/entity/Host.java
@@ -48,12 +48,12 @@ public class Host {
     /**
     * 内存大小
     */
-    private Integer memory;
+    private Long memory;
 
     /**
     * 磁盘大小
     */
-    private Integer disk;
+    private Long disk;
 
     /**
     * 网卡信息
